fix(UiUtils): do not exit the process from `exitWithErrorFnGen()`/`reportAndRejectFnGen()`
Previously, the callback returned by `exitWithErrorFnGen()` would exit the process. This made the
method (and other code that relied on it - e.g. `AbstractCli`) less reusable and more difficult to
test.

This commit fixes this, by changing the callback to return a rejection instead. This gives the
control back to the consumer (which makes sense for a utility method). The method was also renamed
to `reportAndRejectFnGen()` to better reflect its new behavior.